+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import Header from "./components/Header";
+import Header from "@/components/Header";
 import "@/app/global.css"
 const inter = Inter({ subsets: ["latin"] });
