+++ src/components/JobCard.tsx
