+++ src/components/Logo.tsx
@@ -1,4 +1,3 @@
-import "../../../public/logobg.svg"
 export default function Logo() {
   return (
 <svg width="56" height="56" viewBox="0 0 56 56" fill="none" xmlns="http://www.w3.org/2000/svg" xlinkHref="http://www.w3.org/1999/xlink">
