+++ src/components/Menu.tsx
@@ -2,8 +2,8 @@
 import { useState, useEffect } from "react";
 import { usePathname } from "next/navigation";
 import Image from "next/image";
-import { getUserImage, getUsserSesion } from "../utils/userInfo";
-import { doLogout } from "../utils/action";
+import { getUserImage, getUsserSesion } from "@/utils/userInfo";
+import { doLogout } from "@/utils/action";
 
 export default function Menu() {
   const pathName = usePathname();
