+++ src/components/Offers.tsx
