+++ src/images/Computer.tsx
