+++ src/images/Economy.tsx
