+++ src/images/SearchInput.tsx
